requirements.txt
pandas
numpy
scikit-learn
matplotlib
opencv-python
Pillow
tqdm
seaborn
tensorflow
transformers
streamlit
albumentations